Improvement to share action to only include message for mail & message actions
