+++ DuckDuckGo/SyncSettingsViewController+SyncDelegate.swift
@@ -230,18 +230,6 @@ extension SyncSettingsViewController: SyncManagementViewModelDelegate {
         navigationController?.pushViewController(controller, animated: true)
     }
 
-    func shareLink(for url: URL, with message: String, from rect: CGRect) {
-        let activityItems: [Any] = [message, url]
-        let activityViewController = UIActivityViewController(activityItems: activityItems, applicationActivities: nil)
-
-        if UIDevice.current.userInterfaceIdiom == .pad {
-            activityViewController.popoverPresentationController?.sourceView = UIApplication.shared.firstKeyWindow
-            activityViewController.popoverPresentationController?.sourceRect = rect
-        }
-
-        present(activityViewController, animated: true, completion: nil)
-    }
-
     func showPreparingSync() {
         let controller = UIHostingController(rootView: PreparingToSyncView())
         navigationController?.present(controller, animated: true)
